# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "api/pressing/pressings/": {
            "get": {
                "tags": [
                    "Pressing"
                ],
                "summary": "Retrieve all users with the role of 'pressing'",
                "operationId": "21dedb00dbfd2d4e9c81761286e51c5d",
                "responses": {
                    "200": {
                        "description": "List of users with the role of 'pressing'"
                    },
                    "404": {
                        "description": "No users found with the role of 'pressing'"
                    }
                }
            }
        },
        "/api/client/pressings/{id}": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Display the specified pressing",
                "description": "Display the specified pressing.",
                "operationId": "getPressingById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the pressing",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pressing data"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are not authorized to access this resource."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pressing not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Pressing not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/pressing/pressings/update/{id}": {
            "put": {
                "tags": [
                    "Pressing"
                ],
                "summary": "Update a pressing user's profile",
                "operationId": "d62cf5e670c33ab34d4b79487cc495e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the pressing user to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pressing user object that needs to be updated",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Updated pressing user object"
                    },
                    "400": {
                        "description": "Validation errors occurred"
                    },
                    "403": {
                        "description": "The user is not authorized to update this pressing user object"
                    },
                    "404": {
                        "description": "The specified pressing user object was not found"
                    }
                }
            }
        },
        "/api/client/pressings/all": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Get a list of active pressings",
                "description": "Get a list of active pressings.",
                "operationId": "getActivePressings",
                "responses": {
                    "200": {
                        "description": "List of active pressings"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are not authorized to access this resource."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}